requirements.txt
asgiref
Django
djangorestframework
djangorestframework_simplejwt
PyJWT
python-decouple
sqlparse
